trader_asking_implemented_bug_in_sorting

+++ EconomicEngine/Food.cpp
@@ -2,4 +2,4 @@
 
 Food::Food() : Countable(), foodValue(0){}
 
-Food::Food(const int foodValue, const int count = 1) : Countable(count), foodValue(foodValue){}
+Food::Food(const int foodValue, const float count = 1.0f) : Countable(count), foodValue(foodValue){}
